+++ src/main/java/io/jitter/core/shards/ShardsManager.java
@@ -50,7 +50,7 @@ public class ShardsManager implements Managed {
     private final String collection;
     private final String indexPath;
     private Stopper stopper;
-    private final int mu;
+    private final float mu;
     private final String method;
     private final boolean removeDuplicates;
     private final boolean live;
@@ -65,7 +65,7 @@ public class ShardsManager implements Managed {
     private TwitterManager twitterManager;
     private TailyManager tailyManager;
 
-    public ShardsManager(String collection, String indexPath, String stopwords, int mu, String method, boolean removeDuplicates, boolean live, Map<String, Set<String>> topics) {
+    public ShardsManager(String collection, String indexPath, String stopwords, float mu, String method, boolean removeDuplicates, boolean live, Map<String, Set<String>> topics) {
         this.collection = collection;
         this.indexPath = indexPath;
         this.mu = mu;
@@ -130,7 +130,7 @@ public class ShardsManager implements Managed {
         return stopper;
     }
 
-    public int getMu() {
+    public float getMu() {
         return mu;
     }
