+++ src/main/java/io/jitter/core/taily/TailyManagerFactory.java
@@ -20,7 +20,7 @@ public class TailyManagerFactory {
     private String stopwords;
 
     @NotNull
-    private int mu;
+    private float mu;
 
     @NotNull
     private int nc;
@@ -61,12 +61,12 @@ public class TailyManagerFactory {
     }
 
     @JsonProperty
-    public int getMu() {
+    public float getMu() {
         return mu;
     }
 
     @JsonProperty
-    public void setMu(int mu) {
+    public void setMu(float mu) {
         this.mu = mu;
     }
